pyproject.toml
[project]
name = "nhs-conditions-scraper"
version = "0.1.0"
description = "NHS Health A-Z web scraper with FastAPI"
requires-python = ">=3.12"
license = "AGPL-3.0-or-later"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "httpx>=0.28",
    "beautifulsoup4>=4.12",
    "markdownify>=0.14",
    "pyyaml>=6.0",
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.30",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.25",
]

[tool.pytest.ini_options]
asyncio_mode = "strict"
